requirements.txt
evdev; sys_platform == 'linux'
rpyc
pydub

# RPi.GPIO
gpiozero
pigpio
python-mpv